Add checks and solved constructor issue about PNetworking plugin
The construct was called 2 times. The first by UE itsef as default in order to use BP pins. The second one was the correct one, when GetUniqueInstance was called for the first time in bp. Because we need a singleton, a dirty flag is added that doesn't allow to call any initialize methods IF the constructor is not called by the function GetUniqueInstance. So UE can initialize the type, even if pointer is discarded by GC later, and the singleton is initialized when we actually want the plugin during runtime.